Move a few things from net_protocol that don't belong there.
Also document the remaining things in net_protocol
